cleanup, move stuff around
